MDL-55899 mod_data: Updated to use user timezone with dates
Updated calls to data_field_date function to use the default user
timezone instead of no timezone when creating a timestamp.  Otherwise
date is saved as midnight GMT, then when the date is rendered in the
date picker the timezone is considered and, if the user is behind GMT,
the date in the date picker will be a day before what is in the
database.