MDL-77000 markdown: Restore handling of special attributes in <code>
Before the markdown text was treated differently. Now because of
clean_text() special markup that before was contained in the <code>
element (e.g. data-* attributes) was whiped out with HTMLPurifier.
Before cleaning up the HTML we preserve the attributes in the
<code> element, to restore them once the HTMLPurifier has done
its job.
The <code> attributes are used by external tools such as coderunner
to do its job properly.